% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{cost}
\alias{cost}
\title{Custo de um resultado ou solução}
\usage{
cost(x, ...)
}
\arguments{
\item{x}{Um \code{\link[=vrp_solve]{vrp_solve()}} resultado ou uma \code{\link[=vrp_solution]{vrp_solution()}}.}

\item{...}{Não usado.}
}
\value{
O custo objetivo (escalar \code{numeric}); \code{Inf} se nenhuma solução viável
foi encontrada.
}
\description{
Custo de um resultado ou solução
}
