% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ils.R
\name{ils_params}
\alias{ils_params}
\title{ILS solver parameters}
\usage{
ils_params(
  num_neighbours = 50L,
  min_perturbations = 1L,
  max_perturbations = 25L,
  history_length = 300L,
  num_iters_no_improvement = 150000L,
  exhaustive_on_best = TRUE
)
}
\arguments{
\item{num_neighbours}{Granular neighbourhood size (activities per
neighbourhood). Default 50, as in PyVRP.}

\item{min_perturbations, max_perturbations}{Range of perturbations per iteration.}

\item{history_length}{Length of the late-acceptance history (> 0). Default
300, as in PyVRP.}

\item{num_iters_no_improvement}{Iterations without improvement before
restarting from the best.}

\item{exhaustive_on_best}{Refine each new best with an exhaustive search?}
}
\value{
A list of parameters.
}
\description{
ILS solver parameters
}
