% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vrpr_model}
\alias{plot.vrpr_model}
\title{Plot a VRP model (depots and clients only)}
\usage{
\method{plot}{vrpr_model}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=vrp_model]{vrp_model()}}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a VRP model (depots and clients only)
}
