% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{summary.vrpr_result}
\alias{summary.vrpr_result}
\title{Resumo de um resultado, em uma linha (tibble)}
\usage{
\method{summary}{vrpr_result}(object, ...)
}
\arguments{
\item{object}{Um \code{\link[=vrp_solve]{vrp_solve()}} resultado.}

\item{...}{Não usado.}
}
\value{
Um tibble de uma linha com custo, viabilidade, nº de rotas, iterações
e tempo.
}
\description{
Resumo de um resultado, em uma linha (tibble)
}
