% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{vrp_random_solution}
\alias{vrp_random_solution}
\title{Gerar uma solução aleatória}
\usage{
vrp_random_solution(problem_data, seed = 42L)
}
\arguments{
\item{problem_data}{Um \code{\link[=vrp_problem_data]{vrp_problem_data()}}.}

\item{seed}{Semente inteira.}
}
\value{
Um objeto \code{vrpr_solution}.
}
\description{
Gerar uma solução aleatória
}
