% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{vrp_solution}
\alias{vrp_solution}
\title{Build a solution from explicit routes}
\usage{
vrp_solution(problem_data, routes)
}
\arguments{
\item{problem_data}{A \code{\link[=vrp_problem_data]{vrp_problem_data()}}.}

\item{routes}{A list of integer vectors; each vector is a route given as
\emph{client numbers} (1..n_clients), in visit order. All routes use the first
vehicle type.}
}
\value{
A \code{vrpr_solution} object.
}
\description{
Build a solution from explicit routes
}
