% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_depot}
\alias{add_depot}
\title{Adicionar um depósito ao modelo}
\usage{
add_depot(model, x, y)
}
\arguments{
\item{model}{Um \code{vrpr_model}.}

\item{x, y}{Coordenadas do depósito.}
}
\value{
O \code{vrpr_model} atualizado.
}
\description{
Adicionar um depósito ao modelo
}
