% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_vehicle_type}
\alias{add_vehicle_type}
\title{Adicionar um tipo de veículo ao modelo}
\usage{
add_vehicle_type(model, num_available, capacity, fixed_cost = 0)
}
\arguments{
\item{model}{Um \code{vrpr_model}.}

\item{num_available}{Quantidade de veículos disponíveis deste tipo.}

\item{capacity}{Capacidade do veículo.}

\item{fixed_cost}{Custo fixo por veículo usado.}
}
\value{
O \code{vrpr_model} atualizado.
}
\description{
Adicionar um tipo de veículo ao modelo
}
