% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ils.R
\name{ils_params}
\alias{ils_params}
\title{Parâmetros do solver ILS}
\usage{
ils_params(
  num_neighbours = 20L,
  min_perturbations = 1L,
  max_perturbations = 25L,
  init_load = 20,
  init_tw = 6,
  init_dist = 6
)
}
\arguments{
\item{num_neighbours}{Tamanho da vizinhança granular (k vizinhos por cliente).}

\item{min_perturbations, max_perturbations}{Faixa de perturbações por iteração.}

\item{init_load, init_tw, init_dist}{Penalidades iniciais.}
}
\value{
Uma lista de parâmetros.
}
\description{
Parâmetros do solver ILS
}
