% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vrpr_model}
\alias{plot.vrpr_model}
\title{Plotar um modelo VRP (apenas depósitos e clientes)}
\usage{
\method{plot}{vrpr_model}(x, ...)
}
\arguments{
\item{x}{Um \code{\link[=vrp_model]{vrp_model()}}.}

\item{...}{Não usado.}
}
\value{
Um objeto \code{ggplot}.
}
\description{
Plotar um modelo VRP (apenas depósitos e clientes)
}
