% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{routes}
\alias{routes}
\title{Rotas de uma solução, em formato longo (tidy)}
\usage{
routes(x, ...)
}
\arguments{
\item{x}{Um \code{\link[=vrp_solution]{vrp_solution()}}.}

\item{...}{Não usado.}
}
\value{
Um tibble com uma linha por visita: \code{route_id}, \code{position}, \code{client},
\code{vehicle_type}.
}
\description{
Rotas de uma solução, em formato longo (tidy)
}
