% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{unvisited_clients}
\alias{unvisited_clients}
\title{Clientes opcionais não visitados}
\usage{
unvisited_clients(x, ...)
}
\arguments{
\item{x}{Um \code{\link[=vrp_solve]{vrp_solve()}} resultado.}

\item{...}{Não usado.}
}
\value{
Vetor inteiro com os números dos clientes (1-based) não visitados.
}
\description{
Em problemas \emph{prize-collecting}, clientes com \code{required = FALSE} podem ficar
de fora se o prêmio não compensar o custo de roteirização.
}
