% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{vrp_random_solution}
\alias{vrp_random_solution}
\title{Generate a random solution}
\usage{
vrp_random_solution(problem_data, seed = 42L)
}
\arguments{
\item{problem_data}{A \code{\link[=vrp_problem_data]{vrp_problem_data()}}.}

\item{seed}{Integer seed.}
}
\value{
A \code{vrpr_solution} object.
}
\description{
Generate a random solution
}
