% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrpr-package.R
\docType{package}
\name{vrpr-package}
\alias{vrpr}
\alias{vrpr-package}
\title{vrpr: Vehicle Routing Problem Solver Built on 'PyVRP'}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A 'tidyverse'-style interface to high-performance vehicle routing problem (VRP) solving. Vendors the C++ core of the 'PyVRP' solver (\url{https://github.com/PyVRP/PyVRP}) and rewires it through 'cpp11', with no 'Python' runtime dependency. Supports the capacitated VRP, time windows, multiple depots, heterogeneous fleets, prize-collecting and multi-trip variants, driven by an iterated local search metaheuristic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/vrpr}
  \item \url{https://strategicprojects.github.io/vrpr/}
  \item Report bugs at \url{https://github.com/StrategicProjects/vrpr/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br} (\href{https://orcid.org/0000-0001-6249-0920}{ORCID})
  \item Carlos Amorim \email{carlos.agaf@ufpe.br} (\href{https://orcid.org/0000-0001-6315-8305}{ORCID})
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br} (\href{https://orcid.org/0000-0002-1216-8674}{ORCID})
  \item Júlia Nascimento Barreto \email{juliabarreto@gd.seplag.pe.gov.br} (\href{https://orcid.org/0009-0004-2851-7770}{ORCID})
}

Other contributors:
\itemize{
  \item Niels Wouda (author of the vendored PyVRP C++ core) [contributor, copyright holder]
  \item Thibaut Vidal (HGS-CVRP, parts of the vendored C++ core) [copyright holder]
  \item ORTEC (HGS-DIMACS, parts of the vendored C++ core) [copyright holder]
}

}
\keyword{internal}
