% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrpr-package.R
\docType{package}
\name{vrpr-package}
\alias{vrpr}
\alias{vrpr-package}
\title{vrpr: Vehicle Routing Problem Solver (PyVRP for R)}
\description{
A high-performance, tidyverse-style interface to vehicle routing problem (VRP) solving. Ports the PyVRP C++ core to R via cpp11, with no Python runtime dependency. Supports capacitated VRP, time windows, multiple depots, heterogeneous fleets and prize-collecting variants, driven by an iterated local search metaheuristic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/vrpr}
  \item Report bugs at \url{https://github.com/StrategicProjects/vrpr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{leite@de.ufpe.br} (TODO: preencher nome do(a) mantenedor(a))

Authors:
\itemize{
  \item First Last \email{leite@de.ufpe.br} (TODO: preencher nome do(a) mantenedor(a))
}

Other contributors:
\itemize{
  \item  PyVRP contributors (Autores do nucleo C++ PyVRP vendorizado (MIT)) [contributor, copyright holder]
}

}
\keyword{internal}
