% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrpr-package.R
\docType{package}
\name{vrpr-package}
\alias{vrpr}
\alias{vrpr-package}
\title{vrpr: Vehicle Routing Problem Solver (PyVRP for R)}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A high-performance, tidyverse-style interface to vehicle routing problem (VRP) solving. Ports the PyVRP C++ core to R via cpp11, with no Python runtime dependency. Supports capacitated VRP, time windows, multiple depots, heterogeneous fleets and prize-collecting variants, driven by an iterated local search metaheuristic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/vrpr}
  \item Report bugs at \url{https://github.com/StrategicProjects/vrpr/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Andre Leite \email{leite@castlab.org}
  \item Marcos Wasilew \email{marcos.wasilew@gmail.com}
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorin \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
}

Other contributors:
\itemize{
  \item Niels Wouda (author of the vendored PyVRP C++ core) [contributor, copyright holder]
  \item Thibaut Vidal (HGS-CVRP, parts of the vendored C++ core) [copyright holder]
  \item ORTEC (HGS-DIMACS, parts of the vendored C++ core) [copyright holder]
}

}
\keyword{internal}
